Adjust tests so the behavior is not confusing
We should not be able to re-share the shared note to another person.
The button should be hidden as it is very confusing (which one to share?
is it our note or the shared note?) Conveying this message is difficult
to the users, so it's better to make it explicit and do not allow users
to re-share notes on a shared note (UI/UX problem).

Perhaps asking them to just copy the URL bar would be much easier.